now using generics to push in the channel control
